move components and add error alert
